+++ src/EtheriT.Coker.Core/Models/Prod_Price.cs
@@ -7,8 +7,8 @@ namespace EtheriT.Coker.Core.Models
     {
 		public long FK_PSId { get; set; }
         public long FK_RId { get; set; }
-        public double? Price { get; set; }
-        public double? Bonus { get; set; }
+        public decimal? Price { get; set; }
+        public int? Bonus { get; set; }
         public Prod_Stock? Prod_Stock { get; set; }
         public Role? Role { get; set; }
         public List<ShoppingCart> ShoppingCarts { get; set; }
